requirements.txt
docopt
python-gnupg
